@@ -4,63 +4,46 @@
   "private": true,
   "homepage": "/",
   "engines": {
-    "node": "16.x",
-    "npm": "8.x"
+    "node": ">=22.0.0",
+    "npm": ">=10.0.0"
   },
   "dependencies": {
-    "@craco/craco": "^6.1.1",
-    "@fortawesome/fontawesome-free": "^6.7.2",
-    "@material-ui/icons": "^4.11.2",
-    "@react-pdf/renderer": "^1.6.13",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.1",
-    "@testing-library/user-event": "^12.2.0",
-    "axios": "^0.21.4",
-    "bootstrap": "^4.6.0",
-    "concurrently": "^8.2.2",
-    "html2canvas": "^1.0.0-rc.7",
-    "jspdf": "^1.4.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.6.8",
-    "react-cool-img": "^1.2.9",
-    "react-dom": "^16.13.1",
-    "react-emoji": "^0.5.0",
-    "react-ga": "^3.3.0",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@fortawesome/fontawesome-free": "^6.5.1",
+    "@mui/icons-material": "^5.15.6",
+    "@mui/material": "^5.15.6",
+    "@reduxjs/toolkit": "^2.1.0",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "ajv": "^8.12.0",
+    "ajv-keywords": "^5.1.0",
+    "axios": "^1.6.5",
+    "bootstrap": "^5.3.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^6.15.5",
-    "react-icons": "^4.12.0",
-    "react-image-file-resizer": "^0.4.2",
-    "react-instantsearch-dom": "^6.8.2",
-    "react-lazy-load-image-component": "^1.5.1",
-    "react-loadingg": "^1.7.2",
-    "react-notify-toast": "^0.5.1",
-    "react-paypal-button-v2": "^2.6.2",
-    "react-pdf": "^5.2.0",
-    "react-redux": "^7.2.9",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
-    "react-scroll-to-bottom": "^4.1.0",
-    "react-search-field": "^1.2.0",
-    "react-shimmer": "^3.1.1",
-    "react-slick": "^0.27.13",
-    "react-spinners": "^0.10.6",
-    "react-to-pdf": "0.0.13",
-    "react-toast-notifications": "^2.4.4",
-    "react-validation": "^3.0.7",
-    "react-window": "^1.8.6",
-    "redux": "^4.2.1",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.2",
+    "react-hook-form": "^7.49.3",
+    "react-icons": "^5.0.1",
+    "react-redux": "^9.1.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.21.3",
+    "react-scripts": "5.0.1",
+    "react-slick": "^0.29.0",
+    "react-spinners": "^0.13.8",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0",
+    "schema-utils": "^4.2.0",
     "slick-carousel": "^1.8.1",
-    "socket.io-client": "^3.0.4",
-    "sweetalert": "^2.1.2",
-    "web-vitals": "^0.2.4",
-    "webpack-cli": "^4.6.0"
+    "socket.io-client": "^4.7.4",
+    "sweetalert2": "^11.10.3",
+    "web-vitals": "^3.5.1"
   },
   "scripts": {
-    "start": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
-    "build": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -83,10 +66,9 @@
     ]
   },
   "devDependencies": {
-    "@tailwindcss/postcss7-compat": "^2.0.3",
-    "autoprefixer": "^9.8.6",
-    "cross-env": "^7.0.3",
-    "postcss": "^7.0.35",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.3"
+    "postcss": "^8.4.33",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-preset-env": "^9.3.0",
+    "tailwindcss": "^3.4.1"
   }
 }
