@@ -1,9 +1,9 @@
 import React, { Suspense } from 'react';
-import { BrowserRouter, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import { Container } from 'react-bootstrap';
 import Header from './components/Header';
 import Footer from './components/Footer';
-import { CommonLoading } from 'react-loadingg';
+import { CircularProgress } from '@mui/material';
 import DoctorListScreen from './screens/DoctorListScreen';
 
 // Lazy load components
@@ -38,71 +38,42 @@ function App() {
       <div className="app-wrapper">
         <Header />
         <main className="main-content">
-          <Suspense fallback={<CommonLoading />}>
-            <Switch>
-              <Route exact path="/" component={HomeScreen} />
-              <Route exact path="/login" component={LoginScreen} />
-              <Route exact path="/register" component={RegisterScreen} />
-              <Route exact path="/doctor/:id" component={DoctorProfileScreen} />
-              <Route exact path="/booking-appointment/:id" component={BookingAppointmentScreen} />
-              <Route exact path="/booking-success" component={BookingSuccessScreen} />
-              <Route exact path="/tests" component={TestsScreen} />
-              <Route exact path="/test/:id" component={TestBookingScreen} />
-              <Route exact path="/my-test-bookings" component={MyTestBookingsScreen} />
-              <Route exact path="/reschedule-success" component={RescheduleSuccessScreen} />
-              <Route exact path="/admin" render={props => (
-                <AdminRoute>
-                  <AdminDashboardScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/dashboard" render={props => (
-                <AdminRoute>
-                  <AdminDashboardScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/tests" render={props => (
-                <AdminRoute>
-                  <AdminTestsScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/bookings" render={props => (
-                <AdminRoute>
-                  <AdminBookingsScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/users" render={props => (
-                <AdminRoute>
-                  <AdminUsersScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/pharmacy" render={props => (
-                <AdminRoute>
-                  <AdminPharmacyScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/medication/:id/edit" render={props => (
-                <AdminRoute>
-                  <AdminPharmacyScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/pharmacy" component={PharmacyScreen} />
-              <Route exact path="/pharmacy/order/:id" component={MedicationOrderScreen} />
-              <Route exact path="/doctors/specialty/:specialty" component={DoctorsBySpecialtyScreen} />
-              <Route exact path="/prescription-order" component={PrescriptionOrderScreen} />
-              <Route exact path="/prescription-success" component={PrescriptionSuccessScreen} />
-              <Route exact path="/prescription-orders" component={PrescriptionOrderListScreen} />
-              <Route exact path="/admin/prescription-orders" render={props => (
-                <AdminRoute>
-                  <AdminPrescriptionOrdersScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/admin/prescription-orders/add" render={props => (
-                <AdminRoute>
-                  <AddPrescriptionOrderScreen {...props} />
-                </AdminRoute>
-              )} />
-              <Route exact path="/doctors" component={DoctorListScreen} />
-            </Switch>
+          <Suspense fallback={<CircularProgress />}>
+            <Routes>
+              <Route path="/" element={<HomeScreen />} />
+              <Route path="/login" element={<LoginScreen />} />
+              <Route path="/register" element={<RegisterScreen />} />
+              <Route path="/doctor/:id" element={<DoctorProfileScreen />} />
+              <Route path="/booking-appointment/:id" element={<BookingAppointmentScreen />} />
+              <Route path="/booking-success" element={<BookingSuccessScreen />} />
+              <Route path="/tests" element={<TestsScreen />} />
+              <Route path="/test/:id" element={<TestBookingScreen />} />
+              <Route path="/my-test-bookings" element={<MyTestBookingsScreen />} />
+              <Route path="/reschedule-success" element={<RescheduleSuccessScreen />} />
+              
+              {/* Admin Routes */}
+              <Route path="/admin" element={<AdminRoute><AdminDashboardScreen /></AdminRoute>} />
+              <Route path="/admin/dashboard" element={<AdminRoute><AdminDashboardScreen /></AdminRoute>} />
+              <Route path="/admin/tests" element={<AdminRoute><AdminTestsScreen /></AdminRoute>} />
+              <Route path="/admin/bookings" element={<AdminRoute><AdminBookingsScreen /></AdminRoute>} />
+              <Route path="/admin/users" element={<AdminRoute><AdminUsersScreen /></AdminRoute>} />
+              <Route path="/admin/pharmacy" element={<AdminRoute><AdminPharmacyScreen /></AdminRoute>} />
+              <Route path="/admin/medication/:id/edit" element={<AdminRoute><AdminPharmacyScreen /></AdminRoute>} />
+              
+              {/* Pharmacy Routes */}
+              <Route path="/pharmacy" element={<PharmacyScreen />} />
+              <Route path="/pharmacy/order/:id" element={<MedicationOrderScreen />} />
+              <Route path="/doctors/specialty/:specialty" element={<DoctorsBySpecialtyScreen />} />
+              <Route path="/prescription-order" element={<PrescriptionOrderScreen />} />
+              <Route path="/prescription-success" element={<PrescriptionSuccessScreen />} />
+              <Route path="/prescription-orders" element={<PrescriptionOrderListScreen />} />
+              <Route path="/admin/prescription-orders" element={<AdminRoute><AdminPrescriptionOrdersScreen /></AdminRoute>} />
+              <Route path="/admin/prescription-orders/add" element={<AdminRoute><AddPrescriptionOrderScreen /></AdminRoute>} />
+              <Route path="/doctors" element={<DoctorListScreen />} />
+              
+              {/* Catch all route */}
+              <Route path="*" element={<Navigate to="/" replace />} />
+            </Routes>
           </Suspense>
         </main>
         <Footer />
@@ -121,10 +92,9 @@ function App() {
             flex: 1;
             display: flex;
             flex-direction: column;
-            margin-top: 60px; /* Add margin to account for fixed header */
+            margin-top: 60px;
           }
 
-          /* Remove any margin from the last child of main-content to prevent extra space before footer */
           .main-content > *:last-child {
             margin-bottom: 0;
           }
@@ -135,7 +105,6 @@ function App() {
             padding: 0;
           }
 
-          /* Modern scrollbar */
           ::-webkit-scrollbar {
             width: 8px;
             height: 8px;
